app/requirements.txt
# WarDragon Analytics - Python Dependencies
#
# Install with: pip install -r requirements.txt
# For production: pip install --no-cache-dir -r requirements.txt

# Core async framework
asyncio-pool==0.6.0
# Async task pool management for parallel kit polling

# HTTP client for DragonSync API polling
httpx==0.27.0
# Modern async HTTP client with HTTP/2 support
# Used by collector service to fetch from DragonSync APIs

# Database - PostgreSQL/TimescaleDB
sqlalchemy==2.0.25
# SQL toolkit and ORM for Python
# Primary database abstraction layer

psycopg2-binary==2.9.9
# PostgreSQL adapter for Python
# Binary package (no compilation required)
# Alternative: psycopg2 (requires libpq-dev)

asyncpg==0.29.0
# Async PostgreSQL driver
# High-performance alternative to psycopg2 for async operations

alembic==1.13.1
# Database migration tool for SQLAlchemy
# Used for schema versioning and upgrades

# TimescaleDB-specific
timescale==0.1.0
# TimescaleDB helpers for Python
# Simplifies hypertable and continuous aggregate creation

# Configuration management
pyyaml==6.0.1
# YAML parser for kits.yaml and settings.yaml
# Supports loading environment variables in config

python-dotenv==1.0.0
# Load environment variables from .env files
# Used for secrets like DB_PASSWORD, GRAFANA_PASSWORD

# Web framework - FastAPI
fastapi==0.109.0
# Modern async web framework for APIs
# Used for web UI and REST API endpoints

uvicorn[standard]==0.27.0
# ASGI server for FastAPI
# [standard] includes uvloop and httptools for performance

# Web UI templating
jinja2==3.1.3
# Template engine for HTML rendering
# Used for web UI dashboard pages

python-multipart==0.0.6
# Multipart form parsing for FastAPI
# Required for file uploads and form handling

# Authentication (Phase 4 - optional for now)
passlib[bcrypt]==1.7.4
# Password hashing library
# Used when authentication feature is enabled

python-jose[cryptography]==3.3.0
# JWT token handling
# Used for API authentication in Phase 4

# Data processing and export
pandas==2.2.0
# Data analysis library
# Used for CSV export and data aggregation

simplekml==1.3.6
# KML file generation for Google Earth export
# Creates placemarks for drone tracks

geopy==2.4.1
# Geocoding and distance calculations
# Used for geofencing and proximity alerts

# Monitoring and observability
prometheus-client==0.19.0
# Prometheus metrics exporter
# Exposes /metrics endpoint for monitoring

# Logging
python-json-logger==2.0.7
# JSON log formatter
# Structured logging for centralized log aggregation

# Utilities
python-dateutil==2.8.2
# Date/time parsing and manipulation
# Used for query time range handling

pytz==2024.1
# Timezone handling
# Required for UTC conversions and local time display

tenacity==8.2.3
# Retry library with exponential backoff
# Used in collector for resilient kit polling

cachetools==5.3.2
# Caching utilities
# Used for memoization of expensive queries

# Development dependencies (optional)
# Uncomment for development environment
#
# pytest==8.0.0
# # Testing framework
#
# pytest-asyncio==0.23.3
# # Async test support for pytest
#
# pytest-cov==4.1.0
# # Test coverage reporting
#
# black==24.1.1
# # Code formatter
#
# ruff==0.1.14
# # Fast linter (replaces flake8, isort)
#
# mypy==1.8.0
# # Static type checker
#
# ipython==8.20.0
# # Enhanced Python shell for debugging

# Production optimizations
gunicorn==21.2.0
# Production WSGI server (alternative to uvicorn)
# Used for load balancing multiple worker processes

setproctitle==1.3.3
# Set process title for easier monitoring
# Shows descriptive names in ps/top output

# Security
cryptography==42.0.0
# Cryptographic recipes and primitives
# Used for TLS/SSL and secure token generation

certifi==2023.11.17
# Root certificates for TLS verification
# Ensures HTTPS connections are validated

# System monitoring
psutil==5.9.7
# Cross-platform system and process utilities
# Used for collector health metrics

# Optional: Rate limiting (Phase 4)
slowapi==0.1.9
# Rate limiting for FastAPI
# Protects API from abuse

# Optional: WebSocket support for real-time updates
websockets==12.0
# WebSocket protocol implementation
# Used for live dashboard updates (future feature)

# Pin critical dependencies for security and stability
# Update regularly and test before deploying to production
#
# Check for updates: pip list --outdated
# Security audit: pip-audit (requires pip install pip-audit)
